with bg and fg. the ctrl z is fixed

+++ ex1/Commands.cpp
@@ -215,7 +215,7 @@ char** SmallShell::getLastPwd() {
 
 /***************************************** Command methods *****************************************/
 
-Command::Command(const char* cmd_line) {
+Command::Command(const char* cmd_line): last_start_point(time(NULL)) {
   this->cmd_line = new char[strlen(cmd_line)+1];
   strcpy(this->cmd_line, cmd_line);
 }
@@ -453,6 +453,8 @@ void ForegroundCommand::execute() {
     }
   }
 
+  curr_job->resumeProcess();
+
   int status;
   smash.updateCurrentCmd(curr_job->getCommand());
   smash.updateCurrentPid(curr_job->getProcessID());
@@ -463,7 +465,7 @@ void ForegroundCommand::execute() {
   smash.updateCurrentCmd(nullptr);
   smash.updateCurrentPid(0);
 
-  jobs.removeJobById(curr_job->getJobID());
+  jobs.removeFinishedJobs();
 
   _freeArguments(args);
 }
@@ -482,7 +484,7 @@ void BackgroundCommand::execute() {
     _freeArguments(args);
     return;
   }
-
+  
   JobsList::JobEntry* curr_job;
   if(number_of_arguments == 1) {
     JobsList::JobEntry* last = jobs.getLastStoppedJob();
@@ -518,16 +520,19 @@ void BackgroundCommand::execute() {
     }
 
     curr_job = job_to_bg;
+    
   }
 
   cout<<curr_job->getCommand()->getCmdLine()<<" : "<<curr_job->getProcessID()<<endl;
-  if(kill(curr_job->getProcessID(), SIGCONT) == -1) {
-    perror("smash error: kill failed");
-    _freeArguments(args);
-    return;
+  if(curr_job->isStopped()) {
+    if(kill(curr_job->getProcessID(), SIGCONT) == -1) {
+      perror("smash error: kill failed");
+      _freeArguments(args);
+      return;
+    }
   }
-
-  curr_job->turnToBG();
+  //Problems with the timer
+  curr_job->resumeProcess();
   
   _freeArguments(args);
 }
@@ -566,7 +571,6 @@ void ExternalCommand::execute() {
     else {
       smash.updateCurrentPid(p); //
       smash.updateCurrentCmd(this);
-      //wait(nullptr);
       waitpid(p, nullptr, WUNTRACED);
       smash.updateCurrentPid(0);
       smash.updateCurrentCmd(nullptr);
@@ -659,10 +663,14 @@ void JobsList::addJob(Command* cmd, pid_t child_pid, bool isStopped) {
     SmallShell& smash = SmallShell::getInstance();
     int new_job_index = (this->max_id)++;
 
-    cout<< "child_pid" << child_pid;
+    //cout<< "child_pid" << child_pid;
     JobEntry* new_job = new JobEntry(new_job_index, child_pid, isStopped, cmd);
 
     jobs_vector.push_back(new_job);
+
+    if(isStopped) {
+      new_job->stopProcess();
+    }
 }
 
 void JobsList::printJobsList() {
@@ -775,12 +783,30 @@ JobsList::JobEntry* JobsList::getLastStoppedJob(int *jobId) {
   return last_stopped_job;
 }
 
-bool JobsList::jobExistsByPID(pid_t pid) {
+JobsList::JobEntry* JobsList::jobExistsByPID(pid_t pid) {
   for(JobsList::JobEntry* iter:this->jobs_vector) {
     if(iter->getProcessID() == pid) {
-      return true;
+      return iter;
     }
   }
 
-  return false;
+  return nullptr;
+}
+
+
+time_t JobsList::JobEntry::returnDiffTime() { 
+  if(this->is_stopped) {
+    return job_time;
+  }
+  return job_time + difftime(time(NULL), this->cmd->getLastStartPoint()); 
+}
+
+void JobsList::JobEntry::stopProcess() {
+  this->is_stopped = true;
+  job_time = job_time + difftime(time(NULL), this->cmd->getLastStartPoint());
+}
+
+void JobsList::JobEntry::resumeProcess() {
+  this->is_stopped = false;
+  this->cmd->setLastStartPoint();
 }
